use same shared helpers mechanism for SSR compiler
